Adding loading indicator. Change blur image to be a global attribute.
