+++ .vscode/extensions.json
@@ -1,0 +1,9 @@
+{
+  "recommendations": [
+    "AureliaEffect.aurelia",
+    "msjsdiag.debugger-for-chrome",
+    "eg2.tslint",
+    "EditorConfig.EditorConfig",
+    "ricard.PostCSS"
+  ]
+}
