+++ .vscode/settings.json
@@ -1,17 +1,18 @@
 // Place your settings in this file to overwrite default and user settings.
 {
-    "files.exclude": {
-        "**/.git": true,
-        "**/.DS_Store": true,
-        "typings": true,
-        "build": true,
-        "dist": true,
-        "node_modules": true,
-        ".vscode": true
-    },
-    "editor.tabSize": 2,
-    "files.associations": {
-        "/**/*.babelrc": "json",
-        "/**/*.css": "sass-indented"
-    }
+  "typescript.tsdk": "node_modules/typescript/lib",
+  "html.suggest.angular1": false,
+  "html.suggest.ionic": false,
+  "editor.formatOnSave": false,
+  "files.exclude": {
+    "**/.git": true,
+    "**/.DS_Store": true,
+    "node_modules": false,
+    ".vscode": false,
+    "scripts": true
+  },
+  "editor.tabSize": 2,
+  "files.associations": {
+    "/**/*.css": "postcss"
+  }
 }
