+++ aurelia_project/aurelia.json
@@ -117,6 +117,7 @@
           "aurelia-task-queue",
           "aurelia-templating",
           "aurelia-templating-binding",
+          "stackblur-canvas",
           "jquery",
           {
             "name": "bootstrap",
@@ -128,6 +129,14 @@
             "exports": "$",
             "resources": [
               "css/bootstrap.css"
+            ]
+          },
+          {
+            "name": "nprogress",
+            "path": "../node_modules/nprogress",
+            "main": "nprogress",
+            "resources": [
+              "nprogress.css"
             ]
           },
           "text",
