+++ package.json
@@ -14,7 +14,9 @@
     "bluebird": "^3.4.1",
     "bootstrap": "^3.3.7",
     "jquery": "^3.1.1",
+    "nprogress": "^0.2.0",
     "requirejs": "^2.3.2",
+    "stackblur-canvas": "^1.4.0",
     "text": "github:requirejs/text#latest",
     "whatwg-fetch": "^2.0.1"
   },
