+++ src/app.html
@@ -6,6 +6,7 @@
   <nav-bar router.bind="router"></nav-bar>
 
   <div class="page-host">
+    <loading-indicator loading.bind="router.isNavigating || api.isRequesting"></loading-indicator>
     <router-view></router-view>
   </div>
 </template>
