+++ src/resources/index.ts
@@ -1,5 +1,8 @@
 import {FrameworkConfiguration} from 'aurelia-framework';
 
 export function configure(config: FrameworkConfiguration) {
-  //config.globalResources([]);
+  config.globalResources([
+    './attributes/blur-image',
+    './elements/loading-indicator'
+  ]);
 }
