+++ src/users.html
@@ -1,6 +1,4 @@
 <template>
-  <require from="blur-image"></require>
-
   <section class="au-animate">
       <h2>${heading}</h2>
       <div class="row au-stagger">
